// src/store/login/type.ts
/*
 * @Author: your name
 * @Date: 2021-08-15 15:54:55
 * @LastEditTime: 2021-08-15 16:01:15
 * @LastEditors: Please set LastEditors
 * @Description: In User Settings Edit
 * @FilePath: /vue3-practice/src/store/login/type.ts
 */
export interface ILoginState {
  token: string
  userInfo: any
  userMenus: any
}